fixing CI issues
