+++ tests/SCPSpheresElastic/input
@@ -65,6 +65,7 @@ regression.arrhenius.m_ap = 1.45e5
 regression.arrhenius.m_htpb = 1.04e2
 regression.arrhenius.E_ap = 11000.0
 regression.arrhenius.E_htpb = 7500.0
+regression.arrhenius.mob_ap = 1
 
 thermal.on = 1
 thermal.bound = 300.0
@@ -91,22 +92,27 @@ thermal.temp.bc.constant.val.zlo = 0.0
 thermal.temp.bc.constant.val.zhi = 0.0
 
 thermal.hc = 1.0e7
-thermal.mlocal_ap = 1000.0
-thermal.mlocal_htpb = 5000.0
-thermal.mlocal_comb = 0.0
 
 temp.ic.type="constant"
 temp.ic.constant.value=500
 
-pressure.P = 4.0
-pressure.mob_ap = 1
-pressure.a1 = 1.114
-pressure.a2 = 0.46
-pressure.a3 = 2.797
-pressure.b1 = 0.323
-pressure.b2 = 0.42
-pressure.b3 = 0.3225
-pressure.c1 = -0.09906
+chamber.pressure = 4.0
+
+surfacecombustion.type = mesoscale
+surfacecombustion.mesoscale.a1 = 1.114
+surfacecombustion.mesoscale.a2 = 0.46
+surfacecombustion.mesoscale.a3 = 2.797
+surfacecombustion.mesoscale.b1 = 0.323
+surfacecombustion.mesoscale.b2 = 0.42
+surfacecombustion.mesoscale.b3 = 0.3225
+surfacecombustion.mesoscale.c1 = -0.09906
+surfacecombustion.mesoscale.mlocal_ap = 1000.0
+#surfacecombustion.mesoscale.mlocal_htpb = 5000.0
+surfacecombustion.mesoscale.mlocal_comb = 0.0
+
+
+
+
 laser.ic.type = constant
 laser.ic.constant.value = 1.0e6
