+++ tests/SCPThermalSandwich/input
@@ -126,8 +126,8 @@ thermal.temp.bc.constant.val.zhi = 0.0
 thermal.hc = 1.0e7
 
 surfacecombustion.mesoscale.mlocal_ap = 1000.0
-surfacecombustion.mesoscale.mlocal_htpb = 5000.0
-#thermal.mlocal_comb = 0.0
+#surfacecombustion.mesoscale.mlocal_htpb = 5000.0
+surfacecombustion.mesoscale.mlocal_comb = 0.0
 
 ## PRESSURE VARIABLES ######################################################### PRESSURE
